Improved documentation on the testing helpers, and minor refactorings
