+++ Samples/ASP.NET.Specs/Kitchen/for_Kitchen/when_preparing_dish/and_there_are_enough_ingredients.cs
@@ -19,13 +19,13 @@ class and_there_are_enough_ingredients : given.a_kitchen_with_no_prepared_dishes
 
     It should_after_last_operation_only_prepare_one_dish = () => KitchenAfterLastOperation().ShouldHaveEvent<DishPrepared>()
         .CountOf(1)
-        .AtEnd().Where(
+        .AtEnd().AndThat(
             _ => _.Chef.ShouldEqual(chef),
             _ => _.Dish.ShouldEqual(dish));
     
     It should_only_prepare_one_dish = () => KitchenWithAllEvents().ShouldHaveEvent<DishPrepared>()
         .CountOf(1)
-        .AtEnd().Where(
+        .AtEnd().AndThat(
             _ => _.Chef.ShouldEqual(chef),
             _ => _.Dish.ShouldEqual(dish));
