+++ Samples/ASP.NET.Specs/for_DishServer/when_a_dish_has_been_prepared.cs
@@ -15,7 +15,7 @@ class when_a_dish_has_been_prepared : given.the_handler
     
     It the_customer_should_have_eaten_the_dish_once = () => customers.AssertThat(customer_name).ShouldHaveEvent<DishEaten>()
         .CountOf(1)
-        .First().Where(_ => _.Dish.ShouldEqual(@event.Dish));
+        .First().AndThat(_ => _.Dish.ShouldEqual(@event.Dish));
 
     It the_customer_should_have_only_eaten_one_dish = () => customers.AssertThat(customer_name).ShouldHaveNumberOfEvents(1);
