+++ Samples/ASP.NET.Specs/for_DishServer/when_two_dishes_have_been_prepared.cs
@@ -33,17 +33,17 @@ class when_two_dishes_have_been_prepared : given.the_handler
     
     It the_customer_should_have_only_eaten_second_dish_after_last_operation = () => customers.AfterLastOperationOn(customer_name).ShouldHaveEvent<DishEaten>()
         .CountOf(1)
-        .First().Where(_ => _.Dish.ShouldEqual(second_dish.Dish));
+        .First().AndThat(_ => _.Dish.ShouldEqual(second_dish.Dish));
 
     It the_customer_should_have_only_eaten_one_dish_after_last_operation = () => customers.AfterLastOperationOn(customer_name).ShouldHaveNumberOfEvents(1);
 
     It the_customer_should_have_eaten_first_dish_first = () => customers.AssertThat(customer_name).ShouldHaveEvent<DishEaten>()
         .AtBeginning()
-        .Where(_ => _.Dish.ShouldEqual(first_dish.Dish));
+        .AndThat(_ => _.Dish.ShouldEqual(first_dish.Dish));
     
     It the_customer_should_have_eaten_second_dish_last = () => customers.AssertThat(customer_name).ShouldHaveEvent<DishEaten>()
         .AtEnd()
-        .Where(_ => _.Dish.ShouldEqual(second_dish.Dish));
+        .AndThat(_ => _.Dish.ShouldEqual(second_dish.Dish));
 
     It the_customer_should_have_only_eaten_two_dishes = () => customers.AssertThat(customer_name).ShouldHaveNumberOfEvents(2);
