+++ Specifications/Testing/Aggregates/for_AggregateMock/when_performing_operations/on_stateful_aggregate/and_second_operation_fails.cs
@@ -29,5 +29,5 @@ class and_second_operation_fails : given.an_aggregate_of_stateful_aggregate
     It should_assert_that_it_has_the_correct_event = () => aggregate_of.AssertThat(event_source).ShouldHaveEvent<EventCausingStateChange>()
         .CountOf(1)
         .AtEnd()
-        .Where(_ => _.NewState.ShouldEqual(previous_state));
+        .AndThat(_ => _.NewState.ShouldEqual(previous_state));
 }
